验证匹配端；镭射和仿真实测22.04 pcl做原来的帧间匹配慢几十倍；仿真可视化优化

+++ src/robot_localization/src/models/registration/ndt_cpu/Registration.cpp
@@ -89,7 +89,8 @@ void Registration<PointSourceType, PointTargetType>::align(const Eigen::Matrix<f
 
   trans_cloud_.points.resize(source_cloud_->points.size());
 
-  for (int i = 0; i < trans_cloud_.points.size(); i++) {
+  for (int i = 0; i < (int)(trans_cloud_.points.size()); i++) 
+  {
     trans_cloud_.points[i] = source_cloud_->points[i];
   }
